Add files via upload
ID card Detection :  A Python Mini project by PAVAN. S 

The rules of wearing an ID card inside the college is mandated. Students are fined if they violate the rules to revent students from violating the rules but students still try to enter inside the college without an ID card. 
To classify and identif the student without ID card would take 3 securities to do the task as it creates crowd which makes the task a little bit difficult to identify and approach only the students who are not wearing the ID card.

 *ID card Detection allows the Security guards to easily detect , classify students or visitors entering inside the instiution with/without their ID card  and approach only them 
  which will prevent crowding at the entrance gate of the institution.

* Implementing the project inside the institution can help security guards to monitor students who are violating the rule of wearing an ID card. inside the college, 

* ID card Detection enforces the rule of wearing an ID card and helps the security guards to prevent students/ visiters or outsiders form entering into institution.